# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Documentation for Ecommerce-Application",
        "title": "Ecommerce",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:3000",
    "paths": {
...
        "/edit-user": {
            "post": {
                "description": "This Handler enables user to edit his/her details",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "User Edit Profile Details",
                "parameters": [
                    {
                        "description": "name and gender of user",
                        "name": "edit-profile",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "{ \"usercontact\" : \"string\" , \"gender\" : \"string\" }"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/get-user": {
            "get": {
                "description": "This Handler provides all the user information with ID from token",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "User Profile Details",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "description": "Login User with phone number",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "User Login Handler",
                "parameters": [
                    {
                        "description": "phone number of the user",
                        "name": "user-login",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "{ \"usercontact\" : \"string\"}"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/logout": {
            "delete": {
                "description": "This Handler will Log out the user",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "User Logout",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/register": {
            "post": {
                "description": "Registering User with initial details in DB",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "User Register Handler",
                "parameters": [
                    {
                        "description": "name and phone number of the user",
                        "name": "user-register",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "{  \"username\" : \"string\", \"usercontact\" : \"string\"}"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/user/address": {
            "post": {
                "description": "This Handler will set user address",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "User Address",
                "parameters": [
                    {
                        "description": "user address details",
                        "name": "user-address-set",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "{  \"name\" : \"string\",\"street\" : \"string\",\"city\" : \"string\",\"state\" : \"string\",\"postalCode\" : \"string\",\"country\" : \"string\",\"phone\" : \"string\",\"email\" : \"string\",\"addressType\" : \"home/work/default\" }"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/address-get": {
            "get": {
                "description": "This Handler will get user addresses by passing \"addresstype\" param as \"WORK/DEFAULT/HOME\"",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "User Address Retrieve Handler",
                "parameters": [
                    {
                        "type": "string",
                        "description": "address type details",
                        "name": "addresstype",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/verify-otp": {
            "post": {
                "description": "Verify the OTP against the provided phone number",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "User Verify OTP Handler",
                "parameters": [
                    {
                        "description": "phone number and otp of the user",
                        "name": "verify-otp",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "{ \"usercontact\" : \"string\" , \"otp\" : \"string\" }"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}